resilvidas umas besteirinhas

+++ projeto/src/pacotePrincipal/Facade.java
@@ -31,11 +31,11 @@ public class Facade {
 		controllerUsuario.getInfoItem(nome, telefone, nomeItem, atributo);
 	}
 	
-	public void cadastrarEletronico(String nome, String telefone, String nomeItem, int preco, String plataforma) {
+	public void cadastrarEletronico(String nome, String telefone, String nomeItem, double preco, String plataforma) {
 		controllerUsuario.cadastrarEletronico(nome, telefone, nomeItem, preco, plataforma);
 	}
 
-	public void cadastrarJogoTabuleiro(String nome, String telefone, String nomeItem, int preco) {
+	public void cadastrarJogoTabuleiro(String nome, String telefone, String nomeItem, double preco) {
 		controllerUsuario.cadastrarJogoTabuleiro(nome, telefone, nomeItem, preco);
 	}
 
@@ -51,7 +51,7 @@ public class Facade {
 		controllerUsuario.cadastrarBluRayShow(nome, telefone, nomeItem, preco, duracao, numeroFaixas, artista, classificacao);
 	}
 
-	public void cadastrarBluRaySerie(String nome, String telefone, String nomeItem, int preco, String descricao, int duracao, String classificacao, String genero, int temporada) {
+	public void cadastrarBluRaySerie(String nome, String telefone, String nomeItem, double preco, String descricao, int duracao, String classificacao, String genero, int temporada) {
 		controllerUsuario.cadastrarBluRaySerie(nome, telefone, nomeItem, preco, descricao, duracao, classificacao, genero, temporada);
 	}
